Try to separate env files. Separate migrator
