+++ packages/web-app/docker/migrator.dockerfile
@@ -1,13 +1,10 @@
 FROM sophistree/web-app-base AS builder
 
-# Production stage
 FROM node:18-alpine
+
 WORKDIR /app
-COPY --from=builder /app/next.config.js ./
-COPY --from=builder /app/public ./public
-COPY --from=builder /app/.next ./.next
 COPY --from=builder /app/node_modules ./node_modules
 COPY --from=builder /app/package.json ./package.json
+COPY --from=builder /app/prisma ./prisma
 
-EXPOSE 3000
-CMD ["npm", "start"]
+CMD ["npx", "prisma", "migrate", "deploy"]
