+++ packages/web-app/docker/web-app-base.dockerfile
@@ -1,0 +1,8 @@
+FROM node:18-alpine
+
+WORKDIR /app
+COPY package*.json ./
+RUN npm install
+COPY . .
+RUN npx prisma generate
+RUN npm run build
