--- packages/web-app/docker/web-app-entrypoint.sh
@@ -1,10 +1,0 @@
-#!/bin/sh
-set -e
-
-# Run Prisma migrations
-echo "Running Prisma migrations..."
-npx prisma migrate deploy
-
-# Start the application
-echo "Starting the application..."
-exec "$@"
