+++ packages/web-app/src/db/client.ts
@@ -1,5 +1,5 @@
 import { PrismaClient } from "@prisma/client";
-import { SSMClient, GetParameterCommand } from "@aws-sdk/client-ssm";
+import { getPrismaDatabaseUrl } from "./config";
 
 async function makePrismaClient() {
   // Locally, we should provide a DATABASE_URL
@@ -13,13 +13,9 @@ async function makePrismaClient() {
     );
   }
 
-  const password = await getDbPasswordFromParameterStore();
+  const url = await getPrismaDatabaseUrl();
   return new PrismaClient({
-    datasources: {
-      db: {
-        url: `postgresql://sophistree:${password}@sophistree-db:5432/sophistree`,
-      },
-    },
+    datasources: { db: { url } },
   });
 }
 
@@ -36,19 +32,3 @@ export default prismaClientPromise;
 // https://www.prisma.io/docs/orm/more/help-and-troubleshooting/help-articles/nextjs-prisma-client-dev-practices#solution
 if (process.env.NODE_ENV !== "production")
   globalThis.prismaClientPromise = prismaClientPromise;
-
-async function getDbPasswordFromParameterStore() {
-  const ssmClient = new SSMClient({ region: process.env.AWS_REGION });
-  const command = new GetParameterCommand({
-    Name: process.env.DB_PASSWORD_PARAMETER_ARN,
-    WithDecryption: true,
-  });
-
-  try {
-    const response = await ssmClient.send(command);
-    return response.Parameter?.Value;
-  } catch (error) {
-    console.error("Error fetching DB password from SSM:", error);
-    throw error;
-  }
-}
